Implement conv2d_nhwc and depthwise_conv2d_nhwc with textures
* Add evaluate function for different convolutions

* Add depthwise_conv2d_nhwc

* Implement tests for conv2d_nhwc_texture

* Add FP32 accumulator for conv2d_nhwc

* Fix error with types mismatch

* fix minor issues+refactoring

* Add missed nhwc files

+++ tvm/python/tvm/topi/adreno/conv2d_tpack.py
@@ -91,7 +91,7 @@ def compute_conv2d_NCHWc_tpack(Input, Filter, stride, padding, dilation, out_dty
         num_filter_block = 4
     else:
         num_filter_chunk += 1
-    
+
     pad_value = tvm.tir.const(0, Input.dtype)
     def _reorder_data(*indices):
         condition = []
@@ -199,7 +199,7 @@ def compute_conv2d_NCHWc_tpack(Input, Filter, stride, padding, dilation, out_dty
     # leads to the crash in runtime
     # due to this reason we had to use such dummy cast and compute_at to create such intermediate
     # accumulator with local scope
-    dummy_cast = te.compute((batch, num_filter_chunk, out_height_orig, out_width_orig, 4), lambda n,fc,y,x,fb: conv[n,fc,y,x,fb].astype("float16"), tag="dummy_cast")
+    dummy_cast = te.compute((batch, num_filter_chunk, out_height_orig, out_width_orig, 4), lambda n,fc,y,x,fb: conv[n,fc,y,x,fb].astype(out_dtype), tag="dummy_cast")
 
     return te.compute((batch, out_channles, out_height_orig, out_width_orig), lambda n,c,y,x: dummy_cast[n,c // 4,y,x,c % 4], tag="cast_from_acc" + args["accumulator"][-2:])
